move dash leaflet code to frontend directory, rename osm_api module
