+++ frontend/dash_leaflet/app_map.py
@@ -6,7 +6,7 @@ from typing import List
 
 import app_config
 import app_utils
-import pgosm_flex_api
+import frontend.dash_leaflet.osm_api as osm_api
 
 from dash_extensions.javascript import assign
 
@@ -73,7 +73,7 @@ def get_power_grid_overlays(conn: pg.extensions.connection) -> List[dl.Overlay]:
         List[dl.Overlay]: Returns a final list of dl.Overlay components. These are added to the map's LayerControl
         component's children
     """
-    api = pgosm_flex_api.OpenStreetMapDataAPI(conn=conn)
+    api = osm_api.OpenStreetMapDataAPI(conn=conn)
 
     overlays = []
